Create tryit-editor.html
The input seems to be working for now but the run code button doesn't seem to be as it should, instead of a button there is an input prompt or insert bar. so the code that is present there cant be ran with this preview editor code, also still working on the code at the moment.